--- Runtime/Resources/Assets.API.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 20c93c4304bd48f0a4a15a50c2d61d5e
-timeCreated: 1650458667
