--- Runtime/Resources/Assets.Database.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 7e9067e043f44babbfd4c2733e243428
-timeCreated: 1650458684
