+++ Runtime/Utility/Threading/Threading.cs
@@ -2,7 +2,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Threading;
-using Espionage.Engine;
 
 namespace Espionage.Engine
 {
